AAAAAAAAAAAAAAAAAAAAAAAAAAAAA

+++ printf/libft/ft_atoi.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_atoi(const char *str)
+long int	ft_atoi(const char *str)
 {
 	int	i;
 	int	minus;
